+++ include/render/D3D11Render/D3D11Renderer.h
@@ -37,7 +37,7 @@ private:
 
     void initializePipeline(void);
     void initializeDepthAndStencilBuffer(void);
-    void updateEntity(float fDt, ECS::Entity_ID pEntity);
+    void updateEntity(double fDt, ECS::Entity_ID pEntity);
 
     //state variables
     Texture_ID m_currTextID = maxTextureID;
@@ -45,7 +45,7 @@ public:
     D3D11Renderer(ECS::World* world);
     ~D3D11Renderer(void) {}
     bool initialize(void);
-    void update(float dT);
+    void update(double dT);
     void render(void);
     bool shutdown(void);
     inline D3D11Camera* getCamera(void) { return m_pCamera; }
